Create Handler class (#171)
* Created Handler class. Integrated Handler class in StatusHandler.

* Integrated StatusHandler in P2P commands.

* Created EncryptHandler class.

* Created QueryHandler class.

* Created GetDdoHandler and FindDdoHandler.

* Added FeesHandler.

* Added NonceHandler.

* Added download handlers.

* Fix build.

* fix for download handler.

* Combine download URL handler inside DownloadHandler to avoid ReferenceError.

* Removed handler from handlers folder.

* name test.

* debug

* fix import.

* rename back.

* created utils in handlers.

* Fixed imports in tests.

* Include fees handler in aHandler.

* Added status handler to handler. renamed aHandler to handler.

* Added encrypt to handler.

* fix imports.

* Added query handler inside handler.

* fix import in aquarius.

* Added get ddo handler inside handler.

* Added find ddo handler.

* Include findDdo handler inside handler file.

* Fix param.

* Removed handlers folder. Updated imports.

* Big refactor for Handlers.

* Fix imports.

* Added functionality back.

* Fix constant.

* Fixed helper.

* Test only complete flow.

* Fix status test.

* test validation reuse order.

* Added back the tests.

* Removed log. Merged develop into my branch.
