+++ src/components/P2P/handleProtocolCommands.ts
@@ -4,17 +4,16 @@ import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
 import StreamConcat from 'stream-concat'
 // export function handleProtocolCommands (sourceStream:any,sinkStream:any) {
 
-import { handleDownloadURLCommand, handleDownload } from '../core/downloadHandler.js'
-import { PROTOCOL_COMMANDS } from '../../utils/constants.js'
+import { DownloadHandler } from '../core/downloadHandler.js'
+import { PROTOCOL_COMMANDS, HANDLERS_COMMANDS } from '../../utils/constants.js'
 import { P2PCommandResponse } from '../../@types'
 import { P2P_CONSOLE_LOGGER } from './index.js'
-
-import { handleGetDdoCommand, findDDO } from '../core/ddoHandler.js'
-import { getNonce } from '../core/nonceHandler.js'
-import { handleQueryCommand } from '../core/queryHandler.js'
-import { handleStatusCommand } from '../core/statusHandler.js'
-import { handleEncryptCommand } from '../core/encryptHandler.js'
-import { getFees } from '../core/feesHandler.js'
+import { NonceHandler } from '../core/nonceHandler.js'
+import { FeesHandler } from '../core/feesHandler.js'
+import { StatusHandler } from '../core/statusHandler.js'
+import { EncryptHandler } from '../core/encryptHandler.js'
+import { QueryHandler } from '../core/queryHandler.js'
+import { GetDdoHandler, FindDdoHandler } from '../core/ddoHandler.js'
 import { GENERIC_EMOJIS, LOG_LEVELS_STR } from '../../utils/logging/Logger.js'
 
 export class ReadableString extends Readable {
@@ -60,6 +59,22 @@ export async function handleProtocolCommands(connection: any) {
   }
   P2P_CONSOLE_LOGGER.logMessage('Performing task: ' + JSON.stringify(task), true)
 
+  const handlersClasses = [
+    DownloadHandler,
+    EncryptHandler,
+    GetDdoHandler,
+    QueryHandler,
+    NonceHandler,
+    StatusHandler,
+    FindDdoHandler,
+    FeesHandler
+  ]
+  const handlers: Record<string, any> = {}
+  handlersClasses.forEach((HandlerClass, index) => {
+    const handlerInstance = new HandlerClass(this)
+    handlers[HANDLERS_COMMANDS[index]] = handlerInstance
+  })
+
   let response: P2PCommandResponse = null
   try {
     switch (task.command) {
@@ -67,31 +82,28 @@ export async function handleProtocolCommands(connection: any) {
         status = { httpStatus: 200 }
         break
       case PROTOCOL_COMMANDS.DOWNLOAD:
-        response = await handleDownload(this, task)
-        break
-      case PROTOCOL_COMMANDS.DOWNLOAD_URL:
-        response = await handleDownloadURLCommand(this, task)
+        response = await handlers[PROTOCOL_COMMANDS.DOWNLOAD].handle(task)
         break
       case PROTOCOL_COMMANDS.GET_DDO:
-        response = await handleGetDdoCommand(this, task)
+        response = await handlers[PROTOCOL_COMMANDS.GET_DDO].handle(task)
         break
       case PROTOCOL_COMMANDS.QUERY:
-        response = await handleQueryCommand(this, task)
+        response = await handlers[PROTOCOL_COMMANDS.QUERY].handle(task)
         break
       case PROTOCOL_COMMANDS.ENCRYPT:
-        response = await handleEncryptCommand.call(this, task)
+        response = await handlers[PROTOCOL_COMMANDS.ENCRYPT].handle(task)
         break
       case PROTOCOL_COMMANDS.NONCE:
-        response = await getNonce(this, task.address)
+        response = await handlers[PROTOCOL_COMMANDS.NONCE].handle(task)
         break
       case PROTOCOL_COMMANDS.STATUS:
-        response = await handleStatusCommand(task)
+        response = await handlers[PROTOCOL_COMMANDS.STATUS].handle(task)
         break
       case PROTOCOL_COMMANDS.FIND_DDO:
-        response = await findDDO(this, task)
+        response = await handlers[PROTOCOL_COMMANDS.FIND_DDO].handle(task)
         break
       case PROTOCOL_COMMANDS.GET_FEES:
-        response = await getFees(task)
+        response = await handlers[PROTOCOL_COMMANDS.GET_FEES].handle(task)
         break
       default:
         status = { httpStatus: 501, error: 'Unknown command' }
@@ -134,34 +146,50 @@ export async function handleDirectProtocolCommand(message: string, sink: any) {
   let response: P2PCommandResponse = null
 
   P2P_CONSOLE_LOGGER.logMessage('Performing task: ' + JSON.stringify(task), true)
+  const handlersClasses = [
+    DownloadHandler,
+    EncryptHandler,
+    GetDdoHandler,
+    QueryHandler,
+    NonceHandler,
+    StatusHandler,
+    FindDdoHandler,
+    FeesHandler
+  ]
+  const handlers: Record<string, any> = {}
+  handlersClasses.forEach((HandlerClass, index) => {
+    const handlerInstance = new HandlerClass(this)
+    handlers[HANDLERS_COMMANDS[index]] = handlerInstance
+  })
+
   try {
     switch (task.command) {
       case PROTOCOL_COMMANDS.ECHO:
         status = { httpStatus: 200 }
         break
-      case PROTOCOL_COMMANDS.DOWNLOAD_URL:
-        response = await handleDownloadURLCommand(this, task)
+      case PROTOCOL_COMMANDS.DOWNLOAD:
+        response = await handlers[PROTOCOL_COMMANDS.DOWNLOAD].handle(task)
         break
       case PROTOCOL_COMMANDS.GET_DDO:
-        response = await handleGetDdoCommand(this, task)
+        response = await handlers[PROTOCOL_COMMANDS.GET_DDO].handle(task)
         break
       case PROTOCOL_COMMANDS.QUERY:
-        response = await handleQueryCommand(this, task)
+        response = await handlers[PROTOCOL_COMMANDS.QUERY].handle(task)
         break
       case PROTOCOL_COMMANDS.ENCRYPT:
-        response = await handleEncryptCommand.call(this, task)
+        response = await handlers[PROTOCOL_COMMANDS.ENCRYPT].handle(task)
         break
       case PROTOCOL_COMMANDS.NONCE:
-        response = await getNonce(this, task.address)
+        response = await handlers[PROTOCOL_COMMANDS.NONCE].handle(task)
         break
       case PROTOCOL_COMMANDS.STATUS:
-        response = await handleStatusCommand(task)
+        response = await handlers[PROTOCOL_COMMANDS.STATUS].handle(task)
         break
       case PROTOCOL_COMMANDS.FIND_DDO:
-        response = await findDDO(this, task)
+        response = await handlers[PROTOCOL_COMMANDS.FIND_DDO].handle(task)
         break
       case PROTOCOL_COMMANDS.GET_FEES:
-        response = await getFees(task)
+        response = await handlers[PROTOCOL_COMMANDS.GET_FEES].handle(task)
         break
       default:
         status = { httpStatus: 501, error: 'Unknown command' }
