+++ src/components/core/downloadHandler.ts
@@ -1,5 +1,6 @@
-import crypto from 'crypto'
 import { JsonRpcProvider } from 'ethers'
+import { Handler } from './handler.js'
+import { checkNonce, NonceResponse } from './utils/nonceHandler.js'
 import {
   DownloadTask,
   DownloadURLCommand,
@@ -7,226 +8,22 @@ import {
   PROTOCOL_COMMANDS
 } from '../../utils/constants.js'
 import { P2PCommandResponse } from '../../@types/OceanNode.js'
-import { OceanP2P, P2P_CONSOLE_LOGGER } from '../P2P/index.js'
+import { P2P_CONSOLE_LOGGER, OceanP2P } from '../P2P/index.js'
+import { checkFee } from './utils/feesHandler.js'
+import { decrypt } from '../../utils/crypt.js'
+import { FindDdoHandler } from './ddoHandler.js'
+import crypto from 'crypto'
 import * as ethCrypto from 'eth-crypto'
 import { GENERIC_EMOJIS, LOG_LEVELS_STR } from '../../utils/logging/Logger.js'
 import { validateOrderTransaction } from './validateTransaction.js'
-import { checkNonce, NonceResponse } from './nonceHandler.js'
 import { AssetUtils } from '../../utils/asset.js'
 import { Service } from '../../@types/DDO/Service'
-import { findAndFormatDdo } from './ddoHandler.js'
-import { checkFee } from './feesHandler.js'
-import { decrypt } from '../../utils/crypt.js'
 import { ArweaveStorage, IpfsStorage, Storage } from '../../components/storage/index.js'
 import { existsEnvironmentVariable } from '../../utils/index.js'
 import { checkCredentials } from '../../utils/credentials.js'
 export const FILE_ENCRYPTION_ALGORITHM = 'aes-256-cbc'
 
-export async function handleDownload(
-  task: DownloadTask,
-  node: OceanP2P
-): Promise<P2PCommandResponse> {
-  P2P_CONSOLE_LOGGER.logMessage(
-    'Download Request recieved with arguments: ' +
-      task.fileIndex +
-      task.documentId +
-      task.serviceId +
-      task.transferTxId +
-      task.nonce +
-      task.consumerAddress +
-      task.signature,
-    true
-  )
-  // 1. Get the DDO
-  const ddo = await findAndFormatDdo(node, task.documentId)
-
-  if (ddo) {
-    P2P_CONSOLE_LOGGER.logMessage('DDO for asset found: ' + ddo, true)
-  } else {
-    P2P_CONSOLE_LOGGER.logMessage(
-      'No DDO for asset found. Cannot proceed with download.',
-      true
-    )
-    return {
-      stream: null,
-      status: {
-        httpStatus: 500,
-        error: 'No DDO found for asset'
-      }
-    }
-  }
-
-  // 2. Validate ddo and credentials
-  if (!ddo.chainId || !ddo.nftAddress || !ddo.metadata) {
-    P2P_CONSOLE_LOGGER.logMessage('Error: DDO malformed or disabled', true)
-    return {
-      stream: null,
-      status: {
-        httpStatus: 500,
-        error: 'Error: DDO malformed or disabled'
-      }
-    }
-  }
-
-  // check credentials
-  if (ddo.credentials) {
-    const accessGranted = checkCredentials(ddo.credentials, task.consumerAddress)
-    if (!accessGranted) {
-      P2P_CONSOLE_LOGGER.logMessage(`Error: Access to asset ${ddo.id} was denied`, true)
-      return {
-        stream: null,
-        status: {
-          httpStatus: 500,
-          error: `Error: Access to asset ${ddo.id} was denied`
-        }
-      }
-    }
-  }
-
-  // 3. Validate nonce and signature
-  const nonceCheckResult: NonceResponse = await checkNonce(
-    node,
-    task.consumerAddress,
-    parseInt(task.nonce),
-    task.signature,
-    ddo.id
-  )
-
-  if (!nonceCheckResult.valid) {
-    P2P_CONSOLE_LOGGER.logMessage(
-      'Invalid nonce or signature, unable to proceed with download: ' +
-        nonceCheckResult.error,
-      true
-    )
-    return {
-      stream: null,
-      status: {
-        httpStatus: 500,
-        error: nonceCheckResult.error
-      }
-    }
-  }
-
-  // 4. check that the provider fee transaction is valid
-  if (task.feeTx && task.feeData) {
-    let feeValidation
-    try {
-      feeValidation = await checkFee(task.feeTx, task.feeData)
-    } catch (e) {
-      return {
-        stream: null,
-        status: {
-          httpStatus: 500,
-          error: 'ERROR checking fees'
-        }
-      }
-    }
-    if (feeValidation) {
-      // Log the provider fee response for debugging purposes
-      P2P_CONSOLE_LOGGER.logMessage(`Valid provider fee transaction`, true)
-    } else {
-      return {
-        stream: null,
-        status: {
-          httpStatus: 500,
-          error: 'Invalid provider fee transaction'
-        }
-      }
-    }
-  }
-
-  // 5. Call the validateOrderTransaction function to check order transaction
-  const config = node.getConfig()
-  const { rpc } = config.supportedNetworks[ddo.chainId]
-
-  if (!rpc) {
-    P2P_CONSOLE_LOGGER.logMessage(
-      `Cannot proceed with download. RPC not configured for this chain ${ddo.chainId}`,
-      true
-    )
-    return {
-      stream: null,
-      status: {
-        httpStatus: 500,
-        error: `Cannot proceed with download. RPC not configured for this chain ${ddo.chainId}`
-      }
-    }
-  }
-
-  let provider
-  try {
-    provider = new JsonRpcProvider(rpc)
-  } catch (e) {
-    return {
-      stream: null,
-      status: {
-        httpStatus: 500,
-        error: 'JsonRpcProvider ERROR'
-      }
-    }
-  }
-
-  let service: Service = AssetUtils.getServiceById(ddo, task.serviceId)
-  if (!service) service = AssetUtils.getServiceByIndex(ddo, Number(task.serviceId))
-  if (!service) throw new Error('Cannot find service')
-  const paymentValidation = await validateOrderTransaction(
-    task.transferTxId,
-    task.consumerAddress,
-    provider,
-    ddo.nftAddress,
-    service.datatokenAddress,
-    AssetUtils.getServiceIndexById(ddo, task.serviceId),
-    service.timeout
-  )
-
-  if (paymentValidation.isValid) {
-    P2P_CONSOLE_LOGGER.logMessage(
-      `Valid payment transaction. Result: ${paymentValidation.message}`,
-      true
-    )
-  } else {
-    P2P_CONSOLE_LOGGER.logMessage(
-      `Invalid payment transaction: ${paymentValidation.message}`,
-      true
-    )
-    return {
-      stream: null,
-      status: {
-        httpStatus: 500,
-        error: paymentValidation.message
-      }
-    }
-  }
-
-  try {
-    // 6. Decrypt the url
-    const decryptedUrlBytes = await decrypt(
-      Uint8Array.from(Buffer.from(service.files, 'hex')),
-      'ECIES'
-    )
-    // Convert the decrypted bytes back to a string
-    const decryptedFilesString = Buffer.from(decryptedUrlBytes).toString()
-    const decryptedFileArray = JSON.parse(decryptedFilesString)
-    // 7. Proceed to download the file
-    return await handleDownloadURLCommand(node, {
-      command: PROTOCOL_COMMANDS.DOWNLOAD_URL,
-      fileObject: decryptedFileArray.files[task.fileIndex],
-      aes_encrypted_key: task.aes_encrypted_key
-    })
-  } catch (e) {
-    P2P_CONSOLE_LOGGER.logMessage('decryption error' + e, true)
-    return {
-      stream: null,
-      status: {
-        httpStatus: 500,
-        error: 'Failed to decrypt'
-      }
-    }
-  }
-}
-
-// No encryption here yet
-export async function handleDownloadURLCommand(
+export async function handleDownloadUrlCommand(
   node: OceanP2P,
   task: DownloadURLCommand
 ): Promise<P2PCommandResponse> {
@@ -344,6 +141,229 @@ export async function handleDownloadURLCommand(
     return {
       stream: null,
       status: { httpStatus: 501, error: 'Unknown error: ' + err.message }
+    }
+  }
+}
+
+export class DownloadHandler extends Handler {
+  isDownloadCommand(obj: any): obj is DownloadTask {
+    return (
+      typeof obj === 'object' &&
+      obj !== null &&
+      'fileIndex' in obj &&
+      'documentId' in obj &&
+      'serviceId' in obj &&
+      'transferTxId' in obj &&
+      'nonce' in obj &&
+      'consumerAddress' in obj &&
+      'signature' in obj
+    )
+  }
+
+  // No encryption here yet
+
+  async handle(task: any): Promise<P2PCommandResponse> {
+    if (!this.isDownloadCommand(task)) {
+      throw new Error(
+        `Task has nor DownloadCommand type, nor DownloadUrlCommand type. It has ${typeof task}`
+      )
+    }
+    const node = this.getP2PNode()
+    P2P_CONSOLE_LOGGER.logMessage(
+      'Download Request recieved with arguments: ' +
+        task.fileIndex +
+        task.documentId +
+        task.serviceId +
+        task.transferTxId +
+        task.nonce +
+        task.consumerAddress +
+        task.signature,
+      true
+    )
+    // 1. Get the DDO
+    const ddo = await new FindDdoHandler(node).findAndFormatDdo(task.documentId)
+
+    if (ddo) {
+      P2P_CONSOLE_LOGGER.logMessage('DDO for asset found: ' + ddo, true)
+    } else {
+      P2P_CONSOLE_LOGGER.logMessage(
+        'No DDO for asset found. Cannot proceed with download.',
+        true
+      )
+      return {
+        stream: null,
+        status: {
+          httpStatus: 500,
+          error: 'No DDO found for asset'
+        }
+      }
+    }
+
+    // 2. Validate ddo and credentials
+    if (!ddo.chainId || !ddo.nftAddress || !ddo.metadata) {
+      P2P_CONSOLE_LOGGER.logMessage('Error: DDO malformed or disabled', true)
+      return {
+        stream: null,
+        status: {
+          httpStatus: 500,
+          error: 'Error: DDO malformed or disabled'
+        }
+      }
+    }
+
+    // check credentials
+    if (ddo.credentials) {
+      const accessGranted = checkCredentials(ddo.credentials, task.consumerAddress)
+      if (!accessGranted) {
+        P2P_CONSOLE_LOGGER.logMessage(`Error: Access to asset ${ddo.id} was denied`, true)
+        return {
+          stream: null,
+          status: {
+            httpStatus: 500,
+            error: `Error: Access to asset ${ddo.id} was denied`
+          }
+        }
+      }
+    }
+
+    // 3. Validate nonce and signature
+    const nonceCheckResult: NonceResponse = await checkNonce(
+      node,
+      task.consumerAddress,
+      parseInt(task.nonce),
+      task.signature,
+      ddo.id
+    )
+
+    if (!nonceCheckResult.valid) {
+      P2P_CONSOLE_LOGGER.logMessage(
+        'Invalid nonce or signature, unable to proceed with download: ' +
+          nonceCheckResult.error,
+        true
+      )
+      return {
+        stream: null,
+        status: {
+          httpStatus: 500,
+          error: nonceCheckResult.error
+        }
+      }
+    }
+
+    // 4. check that the provider fee transaction is valid
+    if (task.feeTx && task.feeData) {
+      let feeValidation
+      try {
+        feeValidation = await checkFee(task.feeTx, task.feeData)
+      } catch (e) {
+        return {
+          stream: null,
+          status: {
+            httpStatus: 500,
+            error: 'ERROR checking fees'
+          }
+        }
+      }
+      if (feeValidation) {
+        // Log the provider fee response for debugging purposes
+        P2P_CONSOLE_LOGGER.logMessage(`Valid provider fee transaction`, true)
+      } else {
+        return {
+          stream: null,
+          status: {
+            httpStatus: 500,
+            error: 'Invalid provider fee transaction'
+          }
+        }
+      }
+    }
+
+    // 5. Call the validateOrderTransaction function to check order transaction
+    const config = node.getConfig()
+    const { rpc } = config.supportedNetworks[ddo.chainId]
+
+    if (!rpc) {
+      P2P_CONSOLE_LOGGER.logMessage(
+        `Cannot proceed with download. RPC not configured for this chain ${ddo.chainId}`,
+        true
+      )
+      return {
+        stream: null,
+        status: {
+          httpStatus: 500,
+          error: `Cannot proceed with download. RPC not configured for this chain ${ddo.chainId}`
+        }
+      }
+    }
+
+    let provider
+    try {
+      provider = new JsonRpcProvider(rpc)
+    } catch (e) {
+      return {
+        stream: null,
+        status: {
+          httpStatus: 500,
+          error: 'JsonRpcProvider ERROR'
+        }
+      }
+    }
+
+    let service: Service = AssetUtils.getServiceById(ddo, task.serviceId)
+    if (!service) service = AssetUtils.getServiceByIndex(ddo, Number(task.serviceId))
+    if (!service) throw new Error('Cannot find service')
+    const paymentValidation = await validateOrderTransaction(
+      task.transferTxId,
+      task.consumerAddress,
+      provider,
+      ddo.nftAddress,
+      service.datatokenAddress,
+      AssetUtils.getServiceIndexById(ddo, task.serviceId),
+      service.timeout
+    )
+
+    if (paymentValidation.isValid) {
+      P2P_CONSOLE_LOGGER.logMessage(
+        `Valid payment transaction. Result: ${paymentValidation.message}`,
+        true
+      )
+    } else {
+      P2P_CONSOLE_LOGGER.logMessage(
+        `Invalid payment transaction: ${paymentValidation.message}`,
+        true
+      )
+      return {
+        stream: null,
+        status: {
+          httpStatus: 500,
+          error: paymentValidation.message
+        }
+      }
+    }
+
+    try {
+      // 6. Decrypt the url
+      const decryptedUrlBytes = await decrypt(
+        Uint8Array.from(Buffer.from(service.files, 'hex')),
+        'ECIES'
+      )
+      // Convert the decrypted bytes back to a string
+      const decryptedFilesString = Buffer.from(decryptedUrlBytes).toString()
+      const decryptedFileArray = JSON.parse(decryptedFilesString)
+      // 7. Proceed to download the file
+      return await handleDownloadUrlCommand(node, {
+        fileObject: decryptedFileArray.files[task.fileIndex],
+        aes_encrypted_key: task.aes_encrypted_key
+      })
+    } catch (e) {
+      P2P_CONSOLE_LOGGER.logMessage('decryption error' + e, true)
+      return {
+        stream: null,
+        status: {
+          httpStatus: 500,
+          error: 'Failed to decrypt'
+        }
+      }
     }
   }
 }
