+++ src/components/core/statusHandler.ts
@@ -1,121 +1,35 @@
-import os from 'os'
-import {
-  OceanNodeStatus,
-  OceanNodeProvider,
-  OceanNodeIndexer,
-  P2PCommandResponse,
-  StorageTypes
-} from '../../@types/OceanNode.js'
-import {
-  CustomNodeLogger,
-  LOGGER_MODULE_NAMES,
-  LOG_LEVELS_STR,
-  GENERIC_EMOJIS,
-  defaultConsoleTransport,
-  getCustomLoggerForModule
-} from '../../utils/logging/Logger.js'
-import { existsEnvironmentVariable, getConfig } from '../../utils/index.js'
-import { Command, ENVIRONMENT_VARIABLES } from '../../utils/constants.js'
+import { Handler } from './handler.js'
+import { status } from './utils/statusHandler.js'
+import { P2PCommandResponse } from '../../@types/OceanNode.js'
+import { Command } from '../../utils/constants.js'
 import { Readable } from 'stream'
 
-export const STATUS_CONSOLE_LOGGER: CustomNodeLogger = getCustomLoggerForModule(
-  LOGGER_MODULE_NAMES.CORE,
-  LOG_LEVELS_STR.LEVEL_INFO,
-  defaultConsoleTransport
-)
-
-export async function status(nodeId?: string): Promise<OceanNodeStatus> {
-  STATUS_CONSOLE_LOGGER.logMessage('Command status started execution...', true)
-  const config = await getConfig()
-  if (!config) {
-    STATUS_CONSOLE_LOGGER.logMessageWithEmoji(
-      'Config object not found. Cannot proceed with status command.',
-      true,
-      GENERIC_EMOJIS.EMOJI_CROSS_MARK,
-      LOG_LEVELS_STR.LEVEL_ERROR
-    )
-    return
+export class StatusHandler extends Handler {
+  isCommand(obj: any): obj is Command {
+    return typeof obj === 'object' && obj !== null && 'command' in obj
   }
-  const status: OceanNodeStatus = {
-    id: undefined,
-    publicKey: undefined,
-    address: undefined,
-    version: undefined,
-    http: undefined,
-    p2p: undefined,
-    provider: [],
-    indexer: [],
-    supportedStorage: undefined,
-    uptime: process.uptime(),
-    platform: {
-      cpus: os.cpus().length,
-      freemem: os.freemem(),
-      totalmem: os.totalmem(),
-      loadavg: os.loadavg(),
-      arch: os.arch(),
-      machine: os.machine(),
-      platform: os.platform(),
-      release: os.release(),
-      osType: os.type(),
-      osVersion: os.version(),
-      node: process.version
+
+  async handle(task: any): Promise<P2PCommandResponse> {
+    if (!this.isCommand(task)) {
+      throw new Error(`Task has not Command type. It has ${typeof task}`)
     }
-  }
-  if (nodeId && nodeId !== undefined) {
-    status.id = nodeId
-  } else {
-    // get current node ID
-    status.id = config.keys.peerId.toString()
-  }
-
-  const supportedStorageTypes: StorageTypes = {
-    url: true,
-    arwave: existsEnvironmentVariable(ENVIRONMENT_VARIABLES.ARWEAVE_GATEWAY),
-    ipfs: existsEnvironmentVariable(ENVIRONMENT_VARIABLES.IPFS_GATEWAY)
-  }
-
-  status.version = process.env.npm_package_version
-  status.publicKey = Buffer.from(config.keys.publicKey).toString('hex')
-  status.address = config.keys.ethAddress
-  status.http = config.hasHttp
-  status.p2p = config.hasP2P
-  status.supportedStorage = supportedStorageTypes
-
-  if (config.supportedNetworks) {
-    for (const [key, supportedNetwork] of Object.entries(config.supportedNetworks)) {
-      const provider: OceanNodeProvider = {
-        chainId: key,
-        network: supportedNetwork.network
+    try {
+      const statusResult = await status(this.getP2PNode().getConfig(), task.node)
+      if (!statusResult) {
+        return {
+          stream: null,
+          status: { httpStatus: 404, error: 'Status Not Found' }
+        }
       }
-      status.provider.push(provider)
-      const indexer: OceanNodeIndexer = {
-        chainId: key,
-        network: supportedNetwork.network,
-        block: '0'
+      return {
+        stream: Readable.from(JSON.stringify(statusResult)),
+        status: { httpStatus: 200 }
       }
-      status.indexer.push(indexer)
-    }
-  }
-  return status
-}
-
-export async function handleStatusCommand(task: Command): Promise<P2PCommandResponse> {
-  try {
-    const statusResult = await status(task.node)
-    if (!statusResult) {
+    } catch (error) {
       return {
         stream: null,
-        status: { httpStatus: 404, error: 'Config Not Found' }
+        status: { httpStatus: 500, error: 'Unknown error: ' + error.message }
       }
-    }
-    return {
-      stream: Readable.from(JSON.stringify(statusResult)),
-      status: { httpStatus: 200 }
-    }
-  } catch (error) {
-    return {
-      stream: null,
-      status: { httpStatus: 500, error: 'Unknown error: ' + error.message }
     }
   }
 }
